+++ smarts/core/agent.py
@@ -21,8 +21,6 @@ import logging
 import warnings
 from typing import Any, Callable
 
-from smarts.core.observations import Observation
-
 warnings.simplefilter("once")
 
 logger = logging.getLogger(__name__)
@@ -49,7 +47,7 @@ class Agent:
 
         return FunctionAgent()
 
-    def act(self, obs: Observation, **configs):
+    def act(self, obs, **configs):
         """The agent action. See documentation on observations, `AgentSpec`, and `AgentInterface`.
 
         Expects an adapted observation and returns an unadapted action.